backend-datapipe/requirements.txt
Django
djangorestframework
beautifulsoup4
requests
celery
redis
django-celery-beat
django-cors-headers
